requirements.txt
customtkinter
pandas
python-dotenv
github-copilot-sdk